backend/requirements.txt
aniso8601==9.0.1
apispec==4.6.0
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==8.0.1
colorama==0.4.4
cryptography==3.4.7
Flask==2.0.1
flask-apispec==0.11.0
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
idna==2.10
install==1.3.4
isodate==0.6.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
marshmallow==3.12.1
pycparser==2.20
pyparsing==2.4.7
python-dotenv==0.17.1
pytz==2021.1
rdflib==5.0.0
requests==2.25.1
six==1.16.0
SPARQLWrapper==1.8.5
urllib3==1.26.5
webargs==8.0.0
Werkzeug==2.0.1
